@@ -1,10 +1,10 @@
 ﻿using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SpaServices.AngularCli;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using SalaryCounter.Api.Util;
@@ -25,9 +25,9 @@ namespace SalaryCounter.Api
 
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddControllers();
             services.AddMvc()
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_1)
-                .AddJsonOptions(options =>
+                .AddNewtonsoftJson(options =>
                 {
                     options.SerializerSettings.Converters.Add(new StringEnumConverter());
                     options.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
@@ -42,7 +42,8 @@ namespace SalaryCounter.Api
             container.AddSingleton<IAppConfiguration, AppConfiguration>();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        // ReSharper disable once UnusedMember.Global
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
                 app.UseDeveloperExceptionPage();
@@ -52,7 +53,8 @@ namespace SalaryCounter.Api
             app.UseStaticFiles();
             app.UseSpaStaticFiles();
 
-            app.UseMvc();
+            app.UseRouting();
+            app.UseEndpoints(endpoints => endpoints.MapControllers());
             app.UseSpa(spa =>
             {
                 spa.Options.SourcePath = "ClientApp";
